// src/app/features-modules/gestao-de-concurso/api/model/concursoCategoria.ts
/**
 * Concurso.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConcursoCategoriaFase } from './concursoCategoriaFase';
import { EntidadeConcurso } from './entidadeConcurso';
import { ConcursoCategoriaCandidatura } from './concursoCategoriaCandidatura';


export interface ConcursoCategoria { 
    id?: number;
    endidadeConcursoID?: number;
    nome?: string | null;
    descricao?: string | null;
    numVaga?: number;
    dataInicio?: string | null;
    dataFim?: string | null;
    entidadeConcurso?: EntidadeConcurso;
    concursoCategoriaCandidaturas?: Array<ConcursoCategoriaCandidatura> | null;
    concursoCategoriaFases?: Array<ConcursoCategoriaFase> | null;
}