// src/app/features-modules/gestao-de-concurso/api/model/concursoCategoriaFase.ts
/**
 * Concurso.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConcursoCategoria } from './concursoCategoria';
import { Fase } from './fase';


export interface ConcursoCategoriaFase { 
    concursoCategoriaID?: number;
    faseID?: number;
    ordem?: number;
    dataInicio?: string | null;
    dataFim?: string | null;
    concursoCategoria?: ConcursoCategoria;
    fase?: Fase;
}